requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt
dj-database-url==1.2.0
psycopg2==2.9.5
python-decouple==3.7
Pillow==9.4.0
gunicorn

# Testing
coverage
pytest==7.2.1
model-bakery==1.10.1
parameterized==0.8.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-subtests==0.9.0
django-filter==22.1